--- local-interchain/interchain/types/default_chains_test.go
@@ -1,12 +1,0 @@
-package types
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
-
-func TestChains(t *testing.T) {
-	require.NoError(t, ChainCosmosHub().SaveJSON("chains/gen-cosmoshub.json"))
-	require.NoError(t, ChainEthereum().SaveJSON("chains/gen-ethereum.json"))
-}
