+++ local-interchain/interchain/types/genesis.go
@@ -27,7 +27,6 @@ type Genesis struct {
 	Accounts []GenesisAccount `json:"accounts,omitempty" yaml:"accounts,omitempty"`
 
 	// A list of commands which run after chains are good to go.
-	// May need to move out of genesis into its own section? Seems silly though.
 	StartupCommands []string `json:"startup_commands,omitempty" yaml:"startup_commands,omitempty"`
 }
